+++ src/cli.ts
@@ -1,0 +1,36 @@
+import * as path from 'path';
+
+import * as fs from './util/fs';
+import SyncDbOptions from './domain/SyncDbOptions';
+import { CONNECTIONS_FILENAME } from './constants';
+import { resolveConnectionsFromEnv } from './config';
+
+/**
+ * Generates connections.sync-db.json file.
+ *
+ * @returns {Promise<void>}
+ */
+async function generateConnection(): Promise<void> {
+  const filePath = path.resolve(process.cwd(), CONNECTIONS_FILENAME);
+
+  const connections = resolveConnectionsFromEnv();
+  const contents = JSON.stringify({ connections });
+
+  await fs.write(filePath, contents);
+
+  process.stdout.write(`Generated file: ${CONNECTIONS_FILENAME}\n`);
+}
+
+/**
+ * Handle the provided CLI flags.
+ *
+ * @param {SyncDbOptions} flags
+ * @returns {Promise<void>}
+ */
+export async function handleFlags(flags: SyncDbOptions): Promise<void> {
+  if (flags['generate-connections']) {
+    await generateConnection();
+
+    process.exit(0);
+  }
+}
